requirements.txt
asgiref==3.5.2
cffi==1.15.0
cycler==0.11.0
Django==3.2.16
django-cors==0.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.5.3
mip==1.14.1
numpy==1.21.6
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
xlrd==2.0.1
gunicorn
whitenoise